+++ packages/parts/src/lib/index.ts
@@ -1,8 +1,6 @@
 export { default as MediaStream } from './record/MediaStream.svelte';
 export { default as Recorder } from './record/Recorder.svelte';
 
-export { default as Invitation } from './contributors/Invitation.svelte';
-
 export { default as Filter } from './helpers/Filter.svelte';
 
 export { default as LatLngDisplay } from './maps/LatLngDisplay.svelte';
