--- packages/parts/src/routes/+page.md
@@ -1,3 +1,0 @@
-# Living Dictionaries Kitbook
-
-Documentation home and prototyping tool for the [Living Dictionaries](https://livingdictionaries.app) web app, built using Kitbook.
