+++ packages/site/src/index.md
@@ -1,3 +1,3 @@
 # Living Dictionaries Kitbook
 
-In transition.
+Documentation home and prototyping tool for the [Living Dictionaries](https://livingdictionaries.app) web app, built using Kitbook.
